Extract int in StartRemotePeers | fixed script error | moved peerProcess from models
