+++ src/main/java/com/cnt5106/p2p/cleanJava
@@ -4,4 +4,4 @@ rm peerProcess.class
 rm FileParser.class
 rm BTLogger.class
 rm MessageHandler.class
-rm models/MessageTypes.class
+rm models/MessageType.class
